common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro Hurtado",email = "pedrohurt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "google-cloud-firestore (>=2.21.0,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "py-automapper (>=2.2.0,<3.0.0)",
    "dependency-injector (>=4.48.1,<5.0.0)",    
    "uvicorn (>=0.35.0,<0.36.0)",    
    "fastapi[standard] (>=0.116.1,<0.117.0)",    
    "inflect (>=7.5.0,<8.0.0)",
]

[tool.poetry]
packages = [{include = "common", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
